requirements.txt
ninja
pybind11

diffusers
einops
opencv-python
numpy
torch
transformers
torchvision
#taming-transformers-rom1504
#ConfigArgParse
#ipdb
omegaconf

#sentencepiece
tqdm

# Mesh Processing
trimesh>=3.23.5
pymeshlab
pygltflib
xatlas
#kornia
#facexlib

# Training
accelerate
#pytorch_lightning
#scikit-learn
#scikit-image

# Demo only
gradio
fastapi>=0.104.0
uvicorn>=0.23.2
rembg
onnxruntime
#gevent
#geventhttpclient

pydantic>=2.4.2
pillow>=10.0.0
requests>=2.31.0
vedo
# hy3dgen dependencies should be installed separately according to project instructions

aiohttp
